refactor: isolate widget theme bridge
